refactor: update test structure and namespaces for improved organization
- Moved AbstractFunctionalTestCase to a new Support namespace to enhance test organization.
- Updated composer.json to include the new Support namespace for autoloading.
- Adjusted test files to reference the new AbstractFunctionalTestCase location.
