+++ Tests/Functional/Service/ProvideEventListServiceTest.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Cru\Psr14EventList\Tests\Functional\Service;
 
 use Cru\Psr14EventList\Service\ProvideEventListService;
-use Cru\Psr14EventList\Tests\Functional\AbstractFunctionalTestCase;
+use Cru\Psr14EventList\Tests\Functional\Support\AbstractFunctionalTestCase;
 use Cru\Psr14EventList\Tests\Functional\Fixtures\Extensions\EventProvider\Event\AbstractFixtureEvent;
 use Cru\Psr14EventList\Tests\Functional\Fixtures\Extensions\EventProvider\Event\ApplicationReadyEvent;
 use Cru\Psr14EventList\Tests\Functional\Fixtures\Extensions\EventProvider\Event\SubFolder\NestedFixtureEvent;
